+++ flint/test/Flint/Persist/DbSpec.hs
@@ -12,7 +12,7 @@ import qualified Blaze.Import.CallGraph as CG
 import Blaze.Import.Binary (openBinary)
 import Blaze.Import.Source.Ghidra (GhidraImporter)
 import qualified Blaze.Persist.Db as Db
-import Blaze.Types.Function (Func, _name)
+import Blaze.Types.Function (FuncRef, funcRefName, funcRefAddress)
 import qualified Blaze.Types.Graph as G
 
 import qualified Data.HashSet as HashSet
@@ -26,7 +26,7 @@ diveLogger = "res/test_bins/Dive_Logger/Dive_Logger.gzf"
 data TestCtx = TestCtx
   { bv :: GhidraImporter
   , store :: CfgStore
-  , allFuncs :: [Func]
+  , allFuncRefs :: [FuncRef]
   , cg :: CallGraph
   } deriving (Generic)
 
@@ -34,26 +34,26 @@ getTestCtx :: IO TestCtx
 getTestCtx = do
   (bv :: GhidraImporter) <- unsafeFromRight <$> openBinary diveLogger
   store <- CfgStore.init Nothing bv
-  allFuncs <- CG.getFunctions bv
-  cg <- getCallGraph bv allFuncs
+  allFuncRefs <- CG.getFunctions bv
+  cg <- getCallGraph bv allFuncRefs
   return $ TestCtx
     { bv = bv
     , store = store
-    , allFuncs = allFuncs
+    , allFuncRefs = allFuncRefs
     , cg = cg
     }
 
 spec :: Spec
 spec = beforeAll getTestCtx . describe "Flint.Persist.Db" $ do
   context "CallGraph" $ do
-    let reduceGraph g = ( sort . fmap (view _name) . HashSet.toList $ G.nodes g
+    let reduceGraph g = ( sort . fmap (\n -> (funcRefName n, funcRefAddress n)) . HashSet.toList $ G.nodes g
                         , sort $ G.edges g
                         )
 
     it "should store and load CallGraph" $ \tctx -> do
       let action = withSystemTempFile "dive_callgraph.flint" $ \fp _ -> do
               conn <- Db.init fp
-              Db.insertCallGraph conn $ tctx ^. #cg
+              Db.insertCallGraph conn (tctx ^. #cg)
               Db.loadCallGraph conn
 
       fmap reduceGraph <$> action `shouldReturn` Just (reduceGraph $ tctx ^. #cg)